// src/app/models/producto.model.ts
export class Producto {
  IdProducto: string | number | string[] | undefined = 0;
  Nombre: string | number | string[] | undefined = '';
  Tipo: string | number | string[] | undefined = '';
  Precio: string | number | string[] | undefined = 0;
  Piezas: string | number | string[] | undefined = 0;
}